monthly-balances Utiliza servicio para la eliminación de entrada
